componentizing navbar
made navbar to be a component in shared folder, disabled eslint Image component check
